Rewrote main tests file
Gives a nice example of how to use the `ItemAccessWrapper` builder and where `BaseItemContainer` comes from
